Move the duplicated getRetryAttemptTimeoutInMs from event-hubs and service-bus to the core-amqp
