+++ sdk/eventhub/event-hubs/src/managementClient.ts
@@ -12,7 +12,8 @@ import {
   RetryConfig,
   RetryOptions,
   RetryOperationType,
-  SharedKeyCredential
+  SharedKeyCredential,
+  getRetryAttemptTimeoutInMs
 } from "@azure/core-amqp";
 import {
   Message,
@@ -26,7 +27,6 @@ import {
 import { ConnectionContext } from "./connectionContext";
 import { LinkEntity } from "./linkEntity";
 import { logger, logErrorStackTrace } from "./log";
-import { getRetryAttemptTimeoutInMs } from "./util/retries";
 import { AbortSignalLike, AbortError } from "@azure/abort-controller";
 /**
  * Describes the runtime information of an Event Hub.
