+++ sdk/servicebus/service-bus/package.json
@@ -86,7 +86,7 @@
   },
   "dependencies": {
     "@azure/abort-controller": "^1.0.0",
-    "@azure/core-amqp": "^1.1.1",
+    "@azure/core-amqp": "^1.1.3",
     "@azure/core-asynciterator-polyfill": "^1.0.0",
     "@azure/core-http": "^1.1.1",
     "@azure/core-tracing": "1.0.0-preview.8",
