+++ sdk/servicebus/service-bus/src/util/utils.ts
@@ -7,7 +7,7 @@ import { generate_uuid } from "rhea-promise";
 import isBuffer from "is-buffer";
 import { Buffer } from "buffer";
 import * as Constants from "../util/constants";
-import { Constants as CoreAMQPConstants, RetryOptions } from "@azure/core-amqp";
+import { getRetryAttemptTimeoutInMs, RetryOptions } from "@azure/core-amqp";
 
 // This is the only dependency we have on DOM types, so rather than require
 // the DOM lib we can just shim this in.
@@ -40,27 +40,6 @@ export const isNode = typeof navigator === "undefined" && typeof process !== "un
  */
 export function getUniqueName(name: string): string {
   return `${name}-${generate_uuid()}`;
-}
-
-/**
- * @internal
- * @ignore
- *
- * TODO: I think this is duplicated from core-amqp and should be du-duped, but _before_
- * that happens we should question whether it's even a legitimate way of setting the timeout
- * because it just squashes all timeouts beneath 60 seconds to be 60 seconds instead.
- */
-export function getRetryAttemptTimeoutInMs(retryOptions: RetryOptions | undefined): number {
-  const timeoutInMs =
-    retryOptions == undefined ||
-    typeof retryOptions.timeoutInMs !== "number" ||
-    !isFinite(retryOptions.timeoutInMs) ||
-    // TODO: not sure what the justification is for always forcing at least 60 seconds.
-    retryOptions.timeoutInMs < CoreAMQPConstants.defaultOperationTimeoutInMs
-      ? CoreAMQPConstants.defaultOperationTimeoutInMs
-      : retryOptions.timeoutInMs;
-
-  return timeoutInMs;
 }
 
 /**
